check for the variable type for fprintf L190
Hi, 
I came across the error using the newest Matlab version (2019a), using Gurobi solver, on Linux. Shortly the variable sum(row_norms>1e-6) is of a sparse double type, which causes the fprintf to exit with error (as it expect double). I added a simple if statement that checks if the sum(row_norms>1e-6) is a sparse and then uses full(sum(row_norms>1e-6)) as an input for fprintf. If it is not sparse then the code is left unchanged.